chore(api-headless-cms): reenable envs & rename a file
